+++ crystal/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "1.0.0"
-__build__ = "012"
-__release__ = "Crystal V1.00 Build 012"
+__build__ = "013"
+__release__ = "Crystal V1.00 Build 013"
 
 __all__ = ["__version__", "__build__", "__release__"]